+++ PhoneNumberKit/Skip/skip.yml
@@ -1,0 +1,3 @@
+skip:
+  mode: native
+  bridging: true
