+++ Sources/PhoneNumberKit/Info.plist
@@ -1,0 +1,1 @@
+../../PhoneNumberKit/Info.plist
